requirements.txt
heic-to-jpg==0.1.9
heif-convert==1.1.1
pillow==10.3.0
pillow_heif==0.16.0
pycparser==2.22
pyproject==1.3.1
wheel==0.43.0